+++ androidx-kt/src/androidTest/java/me/panpf/androidxkt/test/app/ImplTestInterface.kt
@@ -1,0 +1,4 @@
+package me.panpf.androidxkt.test.app
+
+interface ImplTestInterface {
+}