+++ androidx/src/androidTest/java/me/panpf/androidx/test/app/ImplTestInterface.java
@@ -1,0 +1,4 @@
+package me.panpf.androidx.test.app;
+
+public interface ImplTestInterface {
+}